v1.0 b46

+++ app/src/main/java/com/raffler/app/interfaces/UnreadMessageListener.java
@@ -5,5 +5,5 @@ package com.raffler.app.interfaces;
  */
 
 public interface UnreadMessageListener {
-    void onUnreadMessages(String chatId, int count);
+    void onUnreadMessages(String chatId, long count);
 }
